Fixes #1766. switch_statement_A06_t03.dart. Fix type in a cast pattern (#1773)

+++ LanguageFeatures/Patterns/switch_statement_A06_t03.dart
@@ -56,7 +56,7 @@ String testCast(num value) {
   switch (value) {
     case var _ as int when _ == 0:
       return "cast-1";
-    case var _ as double when _ == 42:
+    case var _ as int when _ == 42:
       return "cast-2";
     default:
       return "default";
